Add request numbering to prevent old (irrelevant) requests from affecting the state of the circuit
Consider 2 following situations ("|" for request started, ">" for request finished, "||" for circuit open interval end):

    |--------------> success                  ||
      |------> error, open circuit breaker    ||

    |--------> opened                         ||
      |---------> success                     ||

In such cases, we don't want the successful response for an old request to
close the circuit.

Given that we can number all requests:

    |-----(1)----> success                         ||
      |---(2)--> error, open circuit breaker       ||
             |----(3)----> whatever                ||

We can see that the circuit was tripped open as part of processing
request 2, whereas request 3 was the last one to start processing so
far. Thus, after opening the circuit, requests with numbers smaller
than or equal to 3 should not close the circuit.